# be/cmd/server/docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /api/bills:
    post:
      consumes:
      - application/json
      description: Create bill
      parameters:
      - collectionFormat: csv
        description: Asset ID
        in: formData
        items:
          type: string
        name: assetId
        required: true
        type: array
      - description: Description
        in: formData
        name: description
        type: string
      - description: Paid or Unpaid
        in: formData
        name: statusBill
        required: true
        type: string
      - description: File to upload
        in: formData
        name: file
        type: file
      - description: Image to upload
        in: formData
        name: image
        type: file
      - description: Buyer Name
        in: formData
        name: buyerName
        required: true
        type: string
      - description: Buyer Phone
        in: formData
        name: buyerPhone
        required: true
        type: string
      - description: Buyer Email
        in: formData
        name: buyerEmail
        required: true
        type: string
      - description: BuyerAddress
        in: formData
        name: buyerAddress
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - JWT: []
      summary: Create bill
      tags:
      - Bills
  /api/bills-un-paid/:
    get:
      consumes:
      - application/json
      description: un paid bill
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - JWT: []
      summary: Get un paid bill
      tags:
      - Bills
  /api/bills/{billNumber}:
    get:
      consumes:
      - application/json
      description: Get bill by bill number
      parameters:
      - description: billNumber
        in: path
        name: billNumber
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - JWT: []
      summary: Get bill by bill number
      tags:
      - Bills
    patch:
      consumes:
      - application/json
      description: Update paid bill
      parameters:
      - description: billNumber
        in: path
        name: billNumber
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - JWT: []
      summary: Update paid bill
      tags:
      - Bills
  /api/bills/filter:
    get:
      consumes:
      - application/json
      description: Get all bill have permission
      parameters:
      - in: query
        name: billNumber
        type: string
      - in: query
        name: categoryId
        type: string
      - format: int64
        in: query
        name: companyId
        type: integer
      - in: query
        name: statusBill
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - JWT: []
      summary: Get all bill with filter
      tags:
      - Bills
...
swagger: "2.0"